fix homework 2
