+++ .gitignore
@@ -1,2 +1,2 @@
-
 /.idea/
+/test-output/
