+++ src/test/java/com/aqa/homeworks/Main.java
@@ -1,0 +1,10 @@
+package com.aqa.homeworks;
+
+public class Main {
+    public static void main(String[] args) {
+
+        TestNGRunner.runTests("smoke");
+        TestNGRunner.runTests("criticalPath");
+        TestNGRunner.runTests("extendedTest");
+    }
+}
