+++ src/test/java/com/aqa/homeworks/tests/CalculatorTest.java
@@ -1,4 +1,4 @@
-package com.aqa.homeworks;
+package com.aqa.homeworks.tests;
 
 import com.epam.tat.module4.Calculator;
 import org.testng.ITestResult;
