+++ src/test/java/com/aqa/homeworks/tests/MultCalcTest.java
@@ -1,31 +1,31 @@
-package com.aqa.homeworks;
+package com.aqa.homeworks.tests;
 
 import com.aqa.homeworks.utils.DataSets;
-import com.epam.tat.module4.Calculator;
+import org.apache.commons.math3.analysis.function.Multiply;
 import org.testng.Assert;
 import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class SumCalcTest extends CalculatorTest {
+public class MultCalcTest extends CalculatorTest{
     @BeforeMethod
     public void beforeTestng(ITestResult itestResult) {
         super.beforeTest(itestResult);
     }
+
     @Test(groups = "smoke", dataProvider = "randomLongData", dataProviderClass = DataSets.class)
-    public void sumLongTest(long firstNumber, long secondNumber) {
-        Assert.assertEquals(getCalculator().sum(firstNumber, secondNumber), firstNumber + secondNumber, "Actual result is not equals expected result");
+    public void subLongTest(long firstNumber, long secondNumber) {
+        Assert.assertEquals(getCalculator().mult(firstNumber, secondNumber), Math.multiplyExact(firstNumber,secondNumber), "Actual result is not equals expected result");
     }
 
     @Test(groups = "smoke", dataProvider = "randomDoubleData", dataProviderClass = DataSets.class)
-    public void sumDoubleTest(double firstNumber, double secondNumber) {
-        Assert.assertEquals(getCalculator().sum(firstNumber, secondNumber), firstNumber + secondNumber, "Actual result is not equals expected result");
+    public void subDoubleTest(double firstNumber, double secondNumber) {
+        Assert.assertEquals(getCalculator().mult(firstNumber, secondNumber), new Multiply().value(firstNumber,secondNumber), "Actual result is not equals expected result");
     }
 
     @AfterMethod
     public void afterTest(ITestResult itestResult) {
         super.afterTest(itestResult);
     }
-
 }
